app/requirements.txt
wheel
Flask==1.0.2
flask-caching==1.4.0
flask-talisman==0.5.0
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
flask-accept==0.0.6
CherryPy==16.0.2
psycopg2-binary==2.7.4
click==6.7
wsgi-request-logger==0.4.6
python-dotenv==0.8.2
Authlib==0.8
pyyaml==3.12
simplejson==3.15.0
sqlalchemy==1.2.8
sqlalchemy-utils==0.33.3
sqlalchemy-json==0.2.1
dateparser==0.7.0
loginpass==0.1.1
python-jose[cryptography]==3.0.0
shortuuid==0.5.0
yamlloader==0.5.2
redis==2.10.6
blinker==1.4
passlib==1.7.1
setuptools==40.0.0
python-crontab==2.3.4
